feat: introduce resolve_imports transformer
