+++ macrotype/modules/emit.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 import types
-from collections import defaultdict
-from collections.abc import Callable as ABC_Callable
 from typing import Annotated, Any, Callable, ForwardRef, Iterable, get_args, get_origin
 
 INDENT = "    "
@@ -10,12 +8,6 @@ INDENT = "    "
 import typing as t
 
 from .ir import ClassDecl, Decl, FuncDecl, ModuleDecl, TypeDefDecl, VarDecl
-
-# Mapping of alias module names to their canonical form.
-_MODULE_ALIASES: dict[str, str] = {
-    "pathlib._local": "pathlib",
-    "pathlib._pathlib": "pathlib",
-}
 
 
 def emit_module(mi: ModuleDecl) -> list[str]:
@@ -28,28 +20,6 @@ def emit_module(mi: ModuleDecl) -> list[str]:
     context = mi.obj.__dict__
     name_map = build_name_map(atoms.values(), context)
 
-    # Precompute typing import names for later checks.
-    typing_names = {
-        name_map[id(a)]
-        for a in atoms.values()
-        if getattr(a, "__module__", None) == "typing" or a in {Callable, ABC_Callable}
-    }
-    typing_names.update(_collect_typing_names(mi.members))
-
-    # Collect imports for non-typing external modules.
-    external: dict[str, set[str]] = defaultdict(set)
-    for atom in atoms.values():
-        modname = getattr(atom, "__module__", None)
-        name = name_map.get(id(atom))
-        if not modname or not name:
-            continue
-        if name in typing_names:
-            continue
-        modname = _MODULE_ALIASES.get(modname, modname)
-        if modname in {"builtins", "typing", mi.obj.__name__}:
-            continue
-        external[modname].add(name)
-
     lines: list[str] = []
     for sym in mi.members:
         if not sym.emit:
@@ -59,39 +29,10 @@ def emit_module(mi: ModuleDecl) -> list[str]:
     if lines and lines[-1] == "":
         lines.pop()
 
-    pre: list[str] = []
-    if external:
-        for mod, names in sorted(external.items()):
-            pre.append(f"from {mod} import {', '.join(sorted(names))}")
-        if typing_names:
-            pre.append("")
-        elif lines:
-            pre.append("")
-    if typing_names:
-        pre.append(f"from typing import {', '.join(sorted(typing_names))}")
-        if lines:
-            pre.append("")
+    pre = mi.imports.lines() if mi.imports else []
+    if pre and lines:
+        return pre + [""] + lines
     return pre + lines
-
-
-def _collect_typing_names(symbols: Iterable[Decl]) -> set[str]:
-    names: set[str] = set()
-    for sym in symbols:
-        if not sym.emit:
-            continue
-        for deco in getattr(sym, "decorators", ()):  # collect decorator names from typing
-            base = deco.split("(")[0].split(".")[-1]
-            if base in {"final", "override", "overload", "runtime_checkable"}:
-                names.add(base)
-        match sym:
-            case TypeDefDecl(obj_type=alias):
-                if isinstance(alias, (t.TypeVar, t.ParamSpec, t.TypeVarTuple)):
-                    names.add(type(alias).__name__)
-                elif alias is t.NewType:
-                    names.add(alias.__name__)
-            case ClassDecl(members=members):
-                names.update(_collect_typing_names(members))
-    return names
 
 
 def _add_comment(line: str, comment: str | None) -> str:
